complete ui changes for pipelineDetails
