+++ openmetadata-ui/src/main/resources/ui/src/styles/app.less
@@ -131,6 +131,9 @@
 .h-9 {
   height: 36px;
 }
+.h-100 {
+  height: 400px;
+}
 .h-min-100 {
   min-height: 100vh;
 }
